[기능 구현] 서버 저장소 dev#1 브랜치와 동기화

+++ app/src/main/java/com/hyojung/quest/Activity/LoginActivity.java
@@ -47,7 +47,7 @@ public class LoginActivity extends AppCompatActivity {
         button_adminAccess.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                LoginQuery clientLoginQuery = new LoginQuery(0, "관리자", "관리자");
+                LoginQuery clientLoginQuery = new LoginQuery((long)0, "관리자", "관리자");
                 Intent intent = new Intent(instance, MainActivity.class);
                 intent.putExtra("kakaoID", (long)0);
                 intent.putExtra("kakaoNickName", "관리자");
